Added a registry field for `QuotaConfig`.
A new field called `quota_configs` is added to persist the
quota configurations of the cluster. This replaces the old
`quotas` field which is deprecated and will be removed
in Mesos 2.0.

When users upgrade to Mesos 1.9, `quotas` will be preserved
and recovered and `quota_configs` will be empty. As users
configures new quotas, whether through the new `UPDATE_QUOTA`
call or the deprecated `SET_QUTOA` call, the configured quotas
will be persisted into the `quota_configs` field along with the
`QUOTA_V2` minimum capability. The capability is removed only
if `quota_configs` becomes empty again. If a role already has an
entry in the old `quotas` field, it will be removed from `quotas`.
In other words, once upgraded, `quotas` will still be preserved
and honored, but it will never grow. Instead it will gradually
shrink as the roles' quotas get updated or removed.

Review: https://reviews.apache.org/r/70950