Added ACLs for agent draining APIs.
This adds three ACLs for the three endpoints added as part of the
agent draining feature:
  * DRAIN_AGENT
  * DEACTIVATE_AGENT
  * REACTIVATE_AGENT

All three ACLs are coarse-grained; a principal is either allowed to
perform the action, or unauthorized. However, there is scope to
restrict the action to particular subsets of agents in future.

Review: https://reviews.apache.org/r/70910/