Added master endpoints for agent draining.
This adds three master calls:
  * DRAIN_AGENT
  * DEACTIVATE_AGENT
  * REACTIVATE_AGENT

DRAIN_AGENT starts automated draining of tasks on the specified agent.
When marked for draining, the agent's resources will not be offered
and all the agent's tasks will be gracefully killed.

DEACTIVATE_AGENT stops offers for a specific agent.  This can be used
for manual draining of tasks, but is not limited to draining.
For example, the operator could deactivate an agent prior to
reserving resources.

REACTIVATE_AGENT restarts offers for a specific agent, that had been
drained or deactivated previously.

Review: https://reviews.apache.org/r/70911/