Implemented `UPDATE_QUOTA` operator call.
This patch wires up the master, auth, registar and allocator
pieces for `UPDATE_QUOTA` call.

This enables the master capability `QUOTA_V2`. The capability
implies the quota v2 API is capable of writes (`UPDATE_QUOTA`)
and the master is capable of recovering from V2 quota
(`QuotaConfig`) in registry.

This patch lacks the rescind offer logic. When quota limits
and guarantees are configured, it might be necessary to
rescind offers on the fly to satisfy new guarantees or be
constrained by the new limits. A todo is left and will be
tackled in subsequent patches.

Also enabled test `MasterQuotaTest.RecoverQuotaEmptyCluster`.

Review: https://reviews.apache.org/r/71021