Added best-effort validation for DRAIN_AGENT master call.
This adds two cases which the master cannot deterministically enforce
when using the DRAIN_AGENT call.  The draining feature is currently
incompatible with maintenance schedules and older versions of the
Mesos agent.  However, in both cases, an agent can reregister with
incompatible info after starting to drain.  Also, unreachable agents
do not have enough info to enforce either case.

These are unexpected cases and are not expected to cause any problems
if circumvented.  However, the agent draining feature is also not
expected to work correctly if circumvented.

Review: https://reviews.apache.org/r/71041