Fixed the flaky RoleTest.RolesEndpointContainsConsumedQuota test.
When this test was run as root in CI, the --launcher gets implicitly
set to `linux` whereas it gets implicitly set to `posix` when run
as non-root. When run as `linux`, the use of cgroups causes the
agents to detect each other's containers as orphans and kill them.

The fix for now is to force the launcher to be posix.

Review: https://reviews.apache.org/r/71048/