Added offer rescind logic for limits enforcement.
Outstanding offers of a role (including offers allocated to its
subroles) are rescinded until the sum of its consumed quota and
outstanding offered resources are below the requested limits.

Note, the rescind effort here is best-effort. It is complex and
expensive to rescind accurately, due to (1) the cost of tracking
the correct resource state (e.g. for limits, tracking of the
precise amount of consumed plus offered (with no reservation
overlap, and similarly, for guarantees, aggregation of all roles'
consumption and outstanding offers) (2) the race between the master
and the allocator. In addition, rescinding offers for quota is
mostly about improving a transient state. Once a quota is set,
hopefully with resource churn, the quota will eventually be
enforced. Lastly, once Mesos starts to adopt an optimistic offer
model (MESOS-1607), quota enforcement will happen during admission
control, rendering offer rescind unnecessary. As a result, we cut
some corners here to only make best effort rescinding.

Specifically, for limits enforcement, we rescind outstanding offers
until the role's consumed and offered is below it's limits. Note,
since consumed and offered might overlap (reservations that are being
offered), this approach might lead to some over rescinding. Also, due
to the race between the master and the allocator, we might rescind
less than we should due to pending offers in the master mailbox.

Also added a test.

Review: https://reviews.apache.org/r/71068