Fixed /roles and GET_ROLES to expose all known roles.
Previously, per MESOS-9888 and MESOS-9890, the /roles and GET_ROLES
APIs only exposed roles that had frameworks associated with them
(either because the framework is subscribed to the role, or there
is a framework with allocations to the role) or configured weight
and/or quota.

This approach omits some important cases:

  (1) Roles that have only reservations associated with them.
  (2) Roles that have only a parent relationship to other roles.

This patch exposes a function that returns all "known" roles based
on the criteria we care about:

  (1) Roles with configured weight or quota.
  (2) Roles with reservations.
  (3) Roles with frameworks subscribed or allocated resources.
  (4) Ancestor roles of (1), (2), or (3).

Also, the resource breakdowns are pulled out from the Role struct
and placed in a function that returns the breakdowns for all known
roles. This was done because there is currently not a Role struct
entry for all known roles.

Review: https://reviews.apache.org/r/71073