Added offer rescind logic for guarantees enforcement.
Outstanding offers need to be rescinded as needed to ensure
roles' requested guarantees can be satisfied.

Note, the rescind effort here is best-effort. It is complex and
expensive to rescind accurately, due to (1) the cost of tracking
the correct resource state (e.g. for limits, tracking of the
precise amount of consumed plus offered (with no reservation
overlap, and similarly, for guarantees, aggregation of all roles'
consumption and outstanding offers) (2) the race between the master
and the allocator. In addition, rescinding offers for quota is
mostly about improving a transient state. Once a quota is set,
hopefully with resource churn, the quota will eventually be
enforced. Lastly, once Mesos starts to adopt an optimistic offer
model (MESOS-1607), quota enforcement will happen during admission
control, rendering offer rescind unnecessary. As a result, we cut
some corners here to only make best effort rescinding.

Specifically, for guarantee enforcement, we pessimistically assume
that what seems like "available" resources in the allocator are
all gone. We greedily rescind offers until rescinded plus sum of
current consumed and offered exceed the total guarantees.

Also since rescinded resources and consumed and offered may overlap
(when rescinded contains reservations that are also part of the
consumed), it is possible that we may under rescind. On the other
hand, we also pessimistically assume that there is no available
resources in the cluster. So chances are we are more likely to
over than under rescind.

Also added a test.

Review: https://reviews.apache.org/r/71111