Added `PendingFutureTracker` class for tracking pending futures.
This patch introduces a mechanism for tracking pending futures.
This feature allows detection of hanging operations, which get
stuck on a blocking operation or asynchronously. However, this
feature does not provide any mechanism for tracking pending
promises, because `Promise` objects might not be accessible in
various cases. Thereby, we introduce a new class that can be
used to track pending futures, so it might facilitate debugging
of stuck issues.

Review: https://reviews.apache.org/r/70887/