Enabled the Docker executor to accept kill policy overrides.
This adds a new `killTask()` overload to the Docker executor
and updates the Mesos executor driver to call into that
overload when the loaded executor is the Docker executor.

This allows the executor driver to pass the kill policy
override, when present, into the Docker executor.

Review: https://reviews.apache.org/r/71034/