Added metrics for quota limits.
This is done similarly to to the existing quota guarantee metrics.

However, the logic is adjusted given the presence of bugs in the
existing implementation. In the case that a non-default quota gets
updated to a new value, the existing logic would fail to add the
new metric and the metric exposed to the user would be stuck with
the original value. To resolve this, we now remove the original
metrics and subsequently add the new ones.

Review: https://reviews.apache.org/r/71188