Refactored master draining test setup.
Tests of this feature will generally require a master, agent, framework,
and a single task to be launched at the beginning of the test.
This moves this common code into the test SetUp.

This also changes the `post(...)` helper to return the http::Response
object instead of parsing it.  The response for DRAIN_AGENT calls
does not return an object, so the tests were not checking the
response before.

Review: https://reviews.apache.org/r/71315