Updated quota related endpoints to return quota configurations.
Added quota configuration information (that includes both
guarantees and limits) in V1 GET_QUOTA call and V0 GET "/quota".

To keep backwards compatibility, the infos field which only
includes the guarantees are continue to be filled. An additional
field configs was added.

Also extended an existing test to cover the changes in
the endpoints.

Review: https://reviews.apache.org/r/71159