Backed `MockResourceProvider` by a process.
We were passing callbacks into `MockResourceProvider` to the HTTP
driver. Since the lifecycle of the callbacks and the mock provider were
decoupled and these callbacks were binding the mock provider instance
the code was not safe as written as the driver could invoke the callback
after the provider had been destructed.

This patch makes sure that the callbacks are defered to a process. We
also dispatch a number of other functions which are strongly coupled to
the lifecycle of the provider. We still do not hide the provider away
completely so the provider can be mocked in tests.

Review: https://reviews.apache.org/r/70728/