Renamed cmake parameter for parallel test execution.
The Jenkins setup (which uses `support/docker-build.sh` under the
covers) is parameterized with the  `CONFIGURATION` environment variable.
While in we pass configure-style flags for both autotools and cmake
builds to it in the Jenkins
configuration, the script performs transformations so that
configure-style flags are transformed to cmake-style (replace `_` with
`-`, uppercase flags, replace `--` with `-D`).

We disable parallel test execution in Jenkins by passing
`--disable-parallel-test-execution` which with the transformations in
`support/docker-build.sh` leads to a cmake arg
`-DDISABLE_PARALLEL_TEST_EXECUTION=1`. This patch renames the cmake arg
from a default enabled `ENABLE_PARALLEL_TEST_EXECUTION` to a default
disabled `DISABLE_PARALLEL_TEST_EXECUTION` to support this workflow.

Review: https://reviews.apache.org/r/71232/