Update `disk/xfs` tests for rootfs quotas.
Parameterize the `disk/xfs` isolator tests on whether we
are enforcing quotas on the sandbox or on the container
rootfs. Separating out the rootfs and sandbox cases wth
parameterization helps to ensure that quota is correctly
applied in both cases and tests don't accidentally pass.

Review: https://reviews.apache.org/r/71195/