Cleared filters upon unsuppressing a role.
Per MESOS-9932, removal of a role from the suppression list does
not currently clear filters. This means that schedulers have to
issue a separate explicit REVIVE for the roles they want to remove.

This ensures that filters are cleared upon removing a role from
the suppression list.

Review: https://reviews.apache.org/r/71264