Fixed agent draining when returning from unreachable state.
This logic was missing from the initial implementation of agent
draining.  When an agent became unreachable, and then reregistered
with the master, the master would not properly deactivate or drain
the agent.

This also fixes a potential problem with checking the agent
drain state too early in the case of pending operations.
Operations are not reported to the master until after the agent
reregisters, so agents with the RESOURCE_PROVIDER capability
cannot be considered DRAINED until after the first
UpdateSlaveMessage has arrived.

Review: https://reviews.apache.org/r/71275