Guarded access to possibly destructed resource provider driver.
We provide a method to tear down a resource provider driver which
internally resets the resource provider HTTP driver pointer. We need to
prevent access to the pointed to value once the driver has been torn
down.

This patch switches all internal uses of the the driver to `send` data
to use a public interface method instead. We also make sure that that
public method does not access an invalidated pointed to value.

Review: https://reviews.apache.org/r/71277