Made sure we are tracking ephemeral quota before getting the usage.
If there is no ephemeral disk resource, the `disk/du` isolator
will not track ephemeral disk usage. In that case, we can't index
the paths hash to find the last usage since it doesn't contain the
ephemeral paths. The fix is to check, and omit the ephemeral usage
if it is not being tracked.

Review: https://reviews.apache.org/r/71283/