Stripped metadata and non-scalars from GET_ROLES resources.
This endpoint is meant to expose resource statistics, and only
accidentally exposed the metadata and non-scalar resources. It
does not make sense to expose non-scalar resources in this way.

The `resources` field will be deprecated in favor of a breakdown
between offered, allocated, and reserved resources (similar to
the /roles endpoint).

Review: https://reviews.apache.org/r/71310