Switched on full debug logging and stack trace reporting for Maven.
This is done in attempt to track down the origin of Java build errors
sporadically observed on CIs. The added volume of logs is around 15K
lines, which is still much smaller than that of the logs of the tests.
Developers who perform local build without Java bindings should
not be affected.

Review: https://reviews.apache.org/r/71337/