Gracefully handled duplicated volumes from non-conforming CSI plugins.
If the SLRP uses a plugin that does not conform to the CSI spec and
reports duplicated volumes, the duplicate would be removed.

This is a backport of `43b86da531a889b1c4b1d7ca6acb2eb924ea01e1`. We
are not backporting test changes as the original patch relies on
periodic plugin polling.

Review: https://reviews.apache.org/r/71769