Added a `dist` target to the cmake build.
This patch adds a `dist` target to the cmake build, analogous to the
target provided by the autotools build.

While cmake already provides a `package_source` target to create a
source archive it does not take care of only including relevant files,
but instead adds all files to the archive, e.g., including build or
backup files which should not be part of a release. For that reason this
patch introduces a script which performs a temporary `git clone` from
the checked out git repository and creates the archive from that clean
tree.

This patch also removes a hardcoded list of ignored files which was
by design not exhaustive.

Review: https://reviews.apache.org/r/71240/