Added a `distcheck` target to the cmake build.
This patch adds a `distcheck` target to the cmake build which mimics the
target of the same name provided by the autotools build. `distcheck`
depends on the `dist` target to create a distribution archive and then
ensures that with the distributed sources the `check` targets succeeds.

Review: https://reviews.apache.org/r/71241/