Revert "Updated cpplint to be compatible with Python 3."
This reverts commit 89db66e3df831eaa50fffb4149a3894097505c14.

This patch was necessary when we were running cpplint in the python3
environment used e.g., also for bindings and other scripts. With
pre-commit we have freedom to choose the Python environment needed so we
can undo our adjustments here to stay closer to upstream.

Review: https://reviews.apache.org/r/71208/