Refactored `removeFramework()` in the allocator.
Now allocated or offered resources are tracked in the slave
struct, we no longer need to look up states in the sorter
when removing frameworks. This patch makes the framework
removal logic more simpler by doing this.

This also paves the way to deprecate the framework sorter.

Review: https://reviews.apache.org/r/71547