[FIX] Rate overflow in port_mapping isolator.
The port mapping isolator currently uses uint32_t to store container
rate limits as Bytes/s. The use of this data type is dictated by the current
libnl interface, that the isolator uses to interface with netlink.

The container preparation phase in the isolator involves setting up TX rate
limiting HTB class using tc command that wants rate to be represented as
bits/s. The conversion is performed by a simple multiplication of the
rate by a literal 8. The result of such multiplication is uint32_t that
can hold a maximum value of 4294967295, which is not enough to represent
speeds available on modern 25G links as bits/s.

This change fixes the described overflow issue. However, representing
network rates as 32bit integers still presents a problem on hosts with
more advanced links, e.g. 100G. This additional problem will be addressed
separately with an upgrade to the recently released libnl 3.5.0, that
features an updated interface with uint64_t rates.