Supported destroying UCR container in `PROVISIONING` state.
Previously in MESOS-3736, we made Docker store support pulling same
image simultaneously which is a performance improvement, however it
may cause an issue: If the pulling hangs somehow, all the subsequent
pulling request for the same image will hang as well, and as a result
the container destroy will also hang since destroy has to wait for
provisioning to finish, see MESOS-4985 for details.

So in this patch we removed that performance improvement and made UCR
can destroy the container which is being provisioned, i.e., UCR will
discard the container provisioning and then keep doing the container
destroy. And we also improved Docker fetcher plugin so that when
container provisioning is discarded the `curl` process used to fetch
manifest or blob will be killed immediately.

Review: https://reviews.apache.org/r/71608