SSL Socket: Implemented socket connection and handshake.
This fills in some of the SSL socket implementation,
in particular the constructor, destructor, connect(),
and accept() methods.

Much of the setup and verification is taken verbatim from the
libevent socket implementation.

A change to the poll socket was necessary to prevent the SSL
socket from holding a self-reference indefinitely.

Review: https://reviews.apache.org/r/71665