Added domain socket-related flags to Mesos agent.
Added two new flags 'http_executor_domain_sockets' and
'domain_socket_location' to the agent, along with minimal
logic to pass the relevant values on to executors.

They are used to control whether the agent should
provide a domain socket to HTTP executors and the
location of the socket on the host filesystem, respectively.

Review: https://reviews.apache.org/r/71816