Updated StreamingHttpConnection to allow writing serialized records.
In order to more efficiently write the initial payload of the
master's operator API event stream, we will need to directly
serialize from the in-memory master state. To enable this, this
patch updates the streaming http connection to allow writing an
already serialized record.

Review: https://reviews.apache.org/r/71826