Handled `/api/v1` and /api/v1/executor` over agent executor socket.
This patch wires up the agent executor socket to handle calls for
`/api/v1` and `/api/v1/executor` so executors can use domain sockets to
communicate with the agent; executors use the latter to subscribe with
the agent and the former to e.g., launch containers.

Note that with this patch we now expose the full operator API over the
agent's domain socket.

Review: https://reviews.apache.org/r/71854