Removed outdated check assuming no reservations with disks.
Since `disk` resources can be re-reserved we need to allow for them to
appear in `RESERVE` operations. The validity of the operation and the
resources appearing in them are already check elsewhere and the check
removed here was intended more as a sanity check/avenue to create a
better error message. With re-reservations it is now incorrect.

Review: https://reviews.apache.org/r/71916/