Allowed disabling use of NVML headers.
NVML is distributed under a license which is probably not "free" which
can e.g., be aan obstacle when including Mesos using symbols from it in
certain distributions.

This patch adds a configure flag to disable use of the NVML headers
completely. Note that with this flag disabled it is impossible to use
GPU isolation with the Mesos containerizer.

Review: https://reviews.apache.org/r/71538/