Disabled `DefaultExecutorTest.DomainSockets` on non-Linux platforms.
While the interfaces are available on all platforms, currently domain
sockets for executor communication are only available on Linux. This
patch ifdef's the code away on non-Linux platforms; this is consistent
with how we disable tests on non-Linux platforms elsewhere where we
never use a test filter.

Review: https://reviews.apache.org/r/71987