Updated master::Call::SUBSCRIBE to be served in parallel.
This call is not entirely read-only, unlike the other GET_* v1 master
calls, and therefore it warranted its own patch.

The approach used is to add a post-processing "write" step to the
handler return type. The post-processing step gets executed
synchronously. In order to deal with different potential post-
processing steps, we use a Variant.

Note that SUBSCRIBE cannot asynchronously register the subscriber
after the read-only state is served, because it will miss events
in the interim!

Review: https://reviews.apache.org/r/72019