Decoupled detection of generated executors from Mesos install location.
We previously were detecting executors generated for command tasks by
checking whether their command matched the full path of
`mesos-executor`.  This approach can lead to misdetection if e.g.,
during an upgrade a new installation location is choosen.

This patch adjusts the heuristic by now only relying on the fact that
the executor command should end in `mesos-executor`. In order to cut
down on false positives we now additionally check that the executor name
looks similar to the ones we generate for command tasks.

Review: https://reviews.apache.org/r/72033/