Added support for 64 bit rates in the port_mapping isolator.
Currenty byte rates in the port_mapping isolator are represented by
uint32_t as in libnl 3.3. Because of that the maximum byte rate
supported by the isolator is ~34 Gbit/s (4294967295*8). This limits the
isolator's capability to utilize higher-speed NIC, e.g. 100G.

This patch changes byte rates to uint64_t, which allows us to support
support values up to ~18 Ebit/s (not UINT64_MAX because of the need to
convert bytes to bits for tc in the isolation script).

This change requires support from libnl. We have to bump the minimal
required version to 3.5.0 to get HTB API with 64 bit rate/ceil.