Removed dead code from v1 API serialization changes.
Per MESOS-10026, we replaced the old protobuf object construction
code with a direct serialization approach to improve performance.
This patch removes the old object construction code (I originally
kept it in case we needed it to debug, but at this point I'm more
confident that we don't need it).

Review: https://reviews.apache.org/r/72068