Added logging of tasks and operations during agent drain initiation.
When draining an agent, it's hard to tell which tasks failed to
terminate from the logs. The master prints a count of the tasks
remaining (only as VLOG(1) however), but not the IDs.

This patch adds logging to the initiation of the drain on both the
master and agent, that shows which tasks and operations are present.
This makes it possible to then see which ones did not transition to
a terminal state (with a bit of log analysis effort).

Review: https://reviews.apache.org/r/72124