Disallow reactivating a DRAINING agent.
When reactivating an agent that's in the draining state, the master
erases it from its draining maps, and erases its estimated drain time.
However, it doesn't send any message to the agent, so if the agent is
still draining and waiting for tasks to terminate, it will stay in
that state, ultimately making any tasks that then get launched get
DROPPED due to the agent still being in a draining state.

To avoid this confusing scenario, we disallow reactivating a DRAINING
agent, since that seems to have been the original design intent.

Review: https://reviews.apache.org/r/72125