Updated Docker containerizer by not updating resources for command task.
For command task, its resources will be set when it is launched as a Docker
container by Docker executor, and we do not need to update its resources
afterward since we do not support task resizing. But for the case that a
custom executor launched as a Docker container by Docker containerizer, we
need to update its resources when it launches a new task or an existing task
terminates.

Review: https://reviews.apache.org/r/72401