Fixed discard race in libwinio io operations.
When getting a discard request for an io operation on windows,
a cancellation is requested and when the io operation completes
a check is performed to see if the future had a discard requested.
However, it's possible that the operation completed successfully
or had a failure and we should not be dropping that information
from the caller.

To correctly check whether the cancellation succeded, we need
to check for ERROR_OPERATION_ABORTED.

Review: https://reviews.apache.org/r/72404