Reverted the changes about `REASON_CONTAINER_MEMORY_REQUEST_EXCEEDED`.
The method `MemorySubsystemProcess::oomWaited()` will only be invoked when the
container is OOM killed because it uses more memory than its hard memory limit
(i.e., the task status reason `REASON_CONTAINER_LIMITATION_MEMORY`), it will
NOT be invoked when a burstable container is OOM killed because the agent host
is running out of memory, i.e., we will NOT receive OOM killing notification
via cgroups notification API for this case. So it is not possible for Mesos to
provide a task status reason `REASON_CONTAINER_MEMORY_REQUEST_EXCEEDED` for this
case.

Review: https://reviews.apache.org/r/72442