Removed unecessary os::execlp wrapper.
This wrapper is deleted on windows and just directly calls through
to execlp on POSIX. The existing callsites don't prefix the call
with os:: so we can remove this wrapper for now.

Also important to note is that we probably want to remove the other
exec wrappers since on Windows they don't make much sense (requires
emulating via a subprocess, can't directly exec like on POSIX).

Review: https://reviews.apache.org/r/72456